requirements.txt
numpy
pandas
torch==2.1.0
torchtext==0.16.0
spacy
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
types-PyYAML 
Flask
Flask-Cors
gdown